Milestone 2

See merge request xmalec/pv-168-easy-food!36

+++ src/main/java/cz/fi/muni/pv168/easyfood/ui/action/DeleteAction.java
@@ -2,19 +2,29 @@ package cz.fi.muni.pv168.easyfood.ui.action;
 
 
 import cz.fi.muni.pv168.easyfood.ui.Icons;
+import cz.fi.muni.pv168.easyfood.ui.MainWindow;
 import cz.fi.muni.pv168.easyfood.ui.tab.TabContainer;
 
 import javax.swing.AbstractAction;
+import javax.swing.JLabel;
+import javax.swing.JOptionPane;
+import javax.swing.JPanel;
 import javax.swing.KeyStroke;
 import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 
+import static javax.swing.JOptionPane.OK_OPTION;
+import static javax.swing.JOptionPane.QUESTION_MESSAGE;
+import static javax.swing.JOptionPane.YES_NO_OPTION;
+
 public final class DeleteAction extends AbstractAction {
     private final TabContainer tabContainer;
 
-    public DeleteAction(TabContainer tabContainer) {
+    private final MainWindow mainWindow;
+    public DeleteAction(MainWindow mainWindow, TabContainer tabContainer) {
         super("Delete", Icons.DELETE_ICON);
         setEnabled(false);
+        this.mainWindow = mainWindow;
         this.tabContainer = tabContainer;
         putValue(SHORT_DESCRIPTION, "Deletes selected row/s");
         putValue(MNEMONIC_KEY, KeyEvent.VK_D);
@@ -23,7 +33,21 @@ public final class DeleteAction extends AbstractAction {
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        tabContainer.getSelectedTab().delete();
+        var table = tabContainer.getSelectedTab().getTable();
+        int selectedRows = table.getSelectedRows().length;
+        if (selectedRows < 1) {
+            throw new IllegalStateException("Invalid selected rows count (must be 1): " + selectedRows);
+        }
+
+        JPanel jPanel = new JPanel();
+        jPanel.add(new JLabel("Delete " + selectedRows + ((selectedRows == 1) ? " row?" : " rows?")));
+
+        int result = JOptionPane.showConfirmDialog(table, jPanel, "Delete",
+                YES_NO_OPTION, QUESTION_MESSAGE, null);
+        if (result == OK_OPTION) {
+            tabContainer.getSelectedTab().delete();
+        }
+        mainWindow.updateRecipeCountLabel();
     }
 }
